refactor(fold-map): split fr_fold_map types + backfill fold leg under the 300-LOC cap
The bloat gate caught a real ratchet: the review-fix round pushed
fr_fold_map.py to 336 and backfill_signals.py to 306. Two cohesive
extractions, no behaviour delta, every import path preserved:

- _fr_fold_map_types.py — the defect vocabulary + FoldDefect/FoldMap/
  FoldResolution records and their serialization (no deps, so no cycle).
  fr_fold_map 336 -> 240.
- _backfill_fold.py — the engine's fold leg, mirroring the collector's
  _test_links_fold; carries the active -> removed -> walk rule order.
  backfill_signals 306 -> 289.
